+++ buildpacks/dotnet/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2025-03-10
+
 ### Changed
 
 - The .NET SDK inventory was updated with new download URLs for version 9.0 release artifacts. ([#203](https://github.com/heroku/buildpacks-dotnet/pull/203))
@@ -122,7 +124,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation.
 
-[unreleased]: https://github.com/heroku/buildpacks-dotnet/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-dotnet/compare/v0.3.1...HEAD
+[0.3.1]: https://github.com/heroku/buildpacks-dotnet/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/heroku/buildpacks-dotnet/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/heroku/buildpacks-dotnet/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/heroku/buildpacks-dotnet/compare/v0.2.0...v0.2.1